@@ -1,14 +1,14 @@
 module github.com/zlymeda/go-ble
 
-go 1.24
+go 1.26
 
 require (
 	github.com/mgutz/logxi v0.0.0-20161027140823-aebf8a7d67ab
 	github.com/pkg/errors v0.9.1
 	github.com/raff/goble v0.0.0-20200327175727-d63360dcfd80
 	github.com/tinygo-org/cbgo v0.0.4
-	github.com/urfave/cli v1.22.16
-	golang.org/x/sys v0.32.0
+	github.com/urfave/cli v1.22.17
+	golang.org/x/sys v0.41.0
 )
 
 require (
@@ -17,5 +17,5 @@ require (
 	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/mgutz/ansi v0.0.0-20200706080929-d51e80ef957d // indirect
 	github.com/russross/blackfriday/v2 v2.1.0 // indirect
-	github.com/sirupsen/logrus v1.9.3 // indirect
+	github.com/sirupsen/logrus v1.9.4 // indirect
 )
